+++ example/lib/main.dart
@@ -1,6 +1,13 @@
 import 'package:example/app.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
-Future<void> main() async {
+import 'flavors.dart';
+
+void main() {
+  F.appFlavor = Flavor.values.firstWhere(
+    (element) => element.name == appFlavor,
+  );
+
   runApp(const App());
 }