+++ lib/src/processors/ide/vscode/vscode_launch_processor.dart
@@ -49,7 +49,7 @@ class VSCodeLaunchProcessor extends StringProcessor {
                     '--flavor',
                     flavorName,
                   ],
-                  program: 'lib/main_$flavorName.dart',
+                  program: 'lib/main.dart',
                 ),
               ),
             )